backend/mysite/requirements.txt
# I will verify each of these to see if them are really necessary
asgiref==3.5.2
astroid==2.12.9
autopep8==1.7.0
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
commonmark==0.9.1
defusedxml==0.7.1
dill==0.3.5.1
Django==4.1.1
django-cors-headers==3.13.0
django-js-asset==2.0.0
django-multiselectfield==0.1.12
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Pillow==9.1.1
platformdirs==2.5.2
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.12.0
PyJWT==2.5.0
pylint==2.15.2
pylint-django==2.5.3
pylint-plugin-utils==0.7
PySocks==1.7.1
pytz==2022.1
requests==2.28.1
requests-toolbelt==0.9.1
rich==12.5.1
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
tzdata==2022.1
urllib3==1.26.11
wrapt==1.14.1